@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": "^7.3||^8.0",
+        "php": "^7.4||^8.0",
         "bacon/bacon-qr-code": "^2.0"
     },
     "require-dev": {
