@@ -6,8 +6,7 @@ namespace Endroid\QrCode\Encoding;
 
 final class Encoding implements EncodingInterface
 {
-    /** @var string */
-    private $value;
+    private string $value;
 
     public function __construct(string $value)
     {
