@@ -8,11 +8,8 @@ use Endroid\QrCode\Label\LabelInterface;
 
 class LabelImageData
 {
-    /** @var int */
-    private $width;
-
-    /** @var int */
-    private $height;
+    private int $width;
+    private int $height;
 
     private function __construct(int $width, int $height)
     {
