@@ -8,23 +8,15 @@ use Endroid\QrCode\Logo\LogoInterface;
 
 class LogoImageData
 {
-    /** @var string */
-    private $data;
+    private string $data;
 
     /** @var mixed */
     private $image;
 
-    /** @var string */
-    private $mimeType;
-
-    /** @var int */
-    private $width;
-
-    /** @var int */
-    private $height;
-
-    /** @var bool */
-    private $punchoutBackground;
+    private string $mimeType;
+    private int $width;
+    private int $height;
+    private bool $punchoutBackground;
 
     /** @param mixed $image */
     private function __construct(
