@@ -6,17 +6,10 @@ namespace Endroid\QrCode\Label\Margin;
 
 final class Margin implements MarginInterface
 {
-    /** @var int */
-    private $top;
-
-    /** @var int */
-    private $right;
-
-    /** @var int */
-    private $bottom;
-
-    /** @var int */
-    private $left;
+    private int $top;
+    private int $right;
+    private int $bottom;
+    private int $left;
 
     public function __construct(int $top, int $right, int $bottom, int $left)
     {
