@@ -10,20 +10,14 @@ use Endroid\QrCode\QrCodeInterface;
 
 final class DebugResult extends AbstractResult
 {
-    /** @var QrCodeInterface */
-    private $qrCode;
-
-    /** @var LogoInterface|null */
-    private $logo;
-
-    /** @var LabelInterface|null */
-    private $label;
+    private QrCodeInterface $qrCode;
+    private ?LogoInterface $logo;
+    private ?LabelInterface $label;
 
     /** @var array<mixed> */
-    private $options;
+    private array $options;
 
-    /** @var bool */
-    private $validateResult = false;
+    private bool $validateResult = false;
 
     /** @param array<mixed> $options */
     public function __construct(QrCodeInterface $qrCode, LogoInterface $logo = null, LabelInterface $label = null, array $options = [])
