@@ -95,7 +95,7 @@ final class PdfWriter implements WriterInterface
 
         if ($label instanceof LabelInterface) {
             $fpdf->SetXY($x, $y + $matrix->getOuterSize() + $labelSpace - 25);
-            $fpdf->SetFont('Helvetica', null, $label->getFont()->getSize());
+            $fpdf->SetFont('Helvetica', '', $label->getFont()->getSize());
             $fpdf->Cell($matrix->getOuterSize(), 0, $label->getText(), 0, 0, 'C');
         }
 
@@ -121,8 +121,8 @@ final class PdfWriter implements WriterInterface
             }
         }
 
-        $logoX = $x + $size / 2 - (int) $logoWidth / 2;
-        $logoY = $y + $size / 2 - (int) $logoHeight / 2;
+        $logoX = $x + $size / 2 - $logoWidth / 2;
+        $logoY = $y + $size / 2 - $logoHeight / 2;
 
         $fpdf->Image($logoPath, $logoX, $logoY, $logoWidth, $logoHeight);
     }
